+++ src/libical/icalenums.h
@@ -58,6 +58,7 @@ typedef enum icalcomponent_kind
     ICAL_PARTICIPANT_COMPONENT,
     ICAL_VLOCATION_COMPONENT,
     ICAL_VRESOURCE_COMPONENT,
+    ICAL_IANA_COMPONENT,
     ICAL_NUM_COMPONENT_TYPES /* MUST be last (unless we can put NO_COMP last) */
 } icalcomponent_kind;